src/main/java/pro/sky/java/course2/repository/impl/JavaQuestionRepository.java
package pro.sky.java.course2.repository.impl;

import org.springframework.stereotype.Repository;
import pro.sky.java.course2.exception.QuestionAlreadyExistException;
import pro.sky.java.course2.exception.QuestionNotFoundException;
import pro.sky.java.course2.model.Question;
import pro.sky.java.course2.repository.QuestionRepository;

import javax.annotation.PostConstruct;
import java.util.Collection;
import java.util.Collections;
import java.util.HashSet;
import java.util.Set;

@Repository
public class JavaQuestionRepository implements QuestionRepository {

    private final Set<Question> questions;

    public JavaQuestionRepository() {
        this.questions = new HashSet<>();
    }

    @PostConstruct
    public void init(){
        add(new Question("Вопрос по Java - 1", "Ответ по Java - 1"));
        add(new Question("Вопрос по Java - 2", "Ответ по Java - 2"));
        add(new Question("Вопрос по Java - 3", "Ответ по Java - 3"));
        add(new Question("Вопрос по Java - 4", "Ответ по Java - 4"));
        add(new Question("Вопрос по Java - 5", "Ответ по Java - 5"));

    }

    @Override
    public Question add(Question question) {
        if (questions.contains(question)){
            throw new QuestionAlreadyExistException();
        }
        questions.add(question);
        return question;
    }

    @Override
    public Question add(String question, String answer) {
        return add(new Question(question, answer));
    }

    @Override
    public Question remove(Question question) {
        if (!questions.contains(question)){
            throw new QuestionNotFoundException();
        }
        questions.remove(question);
        return question;
    }

    @Override
    public Collection<Question> getAll() {
        return Collections.unmodifiableCollection(questions);
    }
}
